refactor(codex): move Responses body preparation into its own module
`prepare_body` and `PreparedBody` grew src/adapters/responses/http.rs from 421
to 623 lines, past the 500-line guideline in AGENTS.md. Move them (and their
tests) to a new `responses::body` module: body preparation is turn-scoped and
shared by both the single-credential and account-pool paths, while http.rs is
the per-send transport, so the split follows an existing seam rather than
cutting a new one.

The `content-encoding: zstd` header now rides with the bytes as
`PreparedBody::attach`, so the header and the body it describes cannot be set
independently. No behavior change: http.rs is back to 420 lines, body.rs is
271, and the wire-shape tests move with the code they cover.
