+++ src/adapters/responses/mod.rs
@@ -1,6 +1,7 @@
 pub mod codex_continuation;
 pub mod codex_ws;
 
+mod body;
 mod context;
 mod error;
 mod http;
